Revert "Move Line / Token / Comment to a separate file. (#154)"
This reverts commit ea66172cb3ab87f988f87baea301b1ba1c122b20,
c19105fce3c3e1790096e8aad3ec6457a485bb96, and
0759421cd17fc4f5927b35600aafbb03f1ae2551.
